refactored duplicate code into one component
